Refactor: Control of the Form managed from CategoryForm
Before: The Control section, when we were using useController, was in
the CreateFieldForm component
After: The control is now in the Category form

The changes were made beacause it doesn't matter where the useController
is called, so the better place for putting it is in the main
CategoryForm where all the useForm functions exists
